Added a way to send the dirpath along the args
Added a way to send the dirpath along the args. If you put the pattern "[dirpath]" in the arguments, this modification will make it replace that pattern with the dirpath variable.

This is useful if you use a cmd that is not the default, and you have to supply the current working directory at the arguments (that's the case with Cmder, for example).